Datenanalyse/Scripts/requirements_CNN.txt
absl-py==0.3.0
astor==0.7.1
backcall==0.1.0
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
cloudpickle==0.5.3
colorama==0.3.9
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
entrypoints==0.2.3
future==0.16.0
gast==0.2.0
grpcio==1.13.0
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.3.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
matplotlib2tikz==0.6.17
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.6.0
numpy==1.15.0
pandas==0.23.3
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
plotly==3.1.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pyts==0.7.1
pytz==2018.5
PyWavelets==0.5.2
pywinpty==0.5.4
pyzmq==17.1.0
qtconsole==4.3.1
requests==2.19.1
retrying==1.3.3
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.3.1
